requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
imageio==2.34.0
lazy_loader==0.3
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
tifffile==2024.2.12
torch==1.12.0
torchvision==0.13.0
typing_extensions==4.10.0
urllib3==2.2.1
